+++ src/windows/chef/cleanup.ps1
@@ -1,55 +1,57 @@
 $ErrorActionPreference = 'Stop'
 $ProgressPreference = 'SilentlyContinue'
 
-$app = Get-WmiObject -Class Win32_Product | Where-Object -Property Name -Match "Chef Infra Client"
-$app.Uninstall()
+if ($env:CHEF_KEEP -ne "true") {
+    $app = Get-WmiObject -Class Win32_Product | Where-Object -Property Name -Match "Chef Infra Client"
+    $app.Uninstall()
 
-$userKey = [Microsoft.Win32.Registry]::CurrentUser.OpenSubKey('Environment',$true)
-$userPath = $userKey.GetValue('PATH', [string]::Empty, 'DoNotExpandEnvironmentNames').ToString()
+    $userKey = [Microsoft.Win32.Registry]::CurrentUser.OpenSubKey('Environment',$true)
+    $userPath = $userKey.GetValue('PATH', [string]::Empty, 'DoNotExpandEnvironmentNames').ToString()
 
-$machineKey = [Microsoft.Win32.Registry]::LocalMachine.OpenSubKey('SYSTEM\ControlSet001\Control\Session Manager\Environment\',$true)
-$machinePath = $machineKey.GetValue('PATH', [string]::Empty, 'DoNotExpandEnvironmentNames').ToString()
+    $machineKey = [Microsoft.Win32.Registry]::LocalMachine.OpenSubKey('SYSTEM\ControlSet001\Control\Session Manager\Environment\',$true)
+    $machinePath = $machineKey.GetValue('PATH', [string]::Empty, 'DoNotExpandEnvironmentNames').ToString()
 
-if ($userPath -like "*$env:ChocolateyInstall*") {
-    $newUserPATH = @(
-        $userPath -split [System.IO.Path]::PathSeparator |
-            Where-Object { $_ -and $_ -ne "$env:ChocolateyInstall\bin" }
-    ) -join [System.IO.Path]::PathSeparator
+    if ($userPath -like "*$env:ChocolateyInstall*") {
+        $newUserPATH = @(
+            $userPath -split [System.IO.Path]::PathSeparator |
+                Where-Object { $_ -and $_ -ne "$env:ChocolateyInstall\bin" }
+        ) -join [System.IO.Path]::PathSeparator
 
-    $userKey.SetValue('PATH', $newUserPATH, 'ExpandString')
-}
-
-if ($machinePath -like "*$env:ChocolateyInstall*") {
-    $newMachinePATH = @(
-        $machinePath -split [System.IO.Path]::PathSeparator |
-            Where-Object { $_ -and $_ -ne "$env:ChocolateyInstall\bin" }
-    ) -join [System.IO.Path]::PathSeparator
-
-    $machineKey.SetValue('PATH', $newMachinePATH, 'ExpandString')
-}
-
-$agentService = Get-Service -Name chocolatey-agent -ErrorAction SilentlyContinue
-if ($agentService -and $agentService.Status -eq 'Running') {
-    $agentService.Stop()
-}
-
-Remove-Item -Path $env:ChocolateyInstall -Recurse -Force
-
-'ChocolateyInstall', 'ChocolateyLastPathUpdate' | ForEach-Object {
-    foreach ($scope in 'User', 'Machine') {
-        [Environment]::SetEnvironmentVariable($_, [string]::Empty, $scope)
+        $userKey.SetValue('PATH', $newUserPATH, 'ExpandString')
     }
-}
 
-$machineKey.Close()
-$userKey.Close()
+    if ($machinePath -like "*$env:ChocolateyInstall*") {
+        $newMachinePATH = @(
+            $machinePath -split [System.IO.Path]::PathSeparator |
+                Where-Object { $_ -and $_ -ne "$env:ChocolateyInstall\bin" }
+        ) -join [System.IO.Path]::PathSeparator
 
-if ($env:ChocolateyToolsLocation -and (Test-Path $env:ChocolateyToolsLocation)) {
-    Remove-Item -Path $env:ChocolateyToolsLocation -Recurse -Force
-}
+        $machineKey.SetValue('PATH', $newMachinePATH, 'ExpandString')
+    }
 
-foreach ($scope in 'User', 'Machine') {
-    [Environment]::SetEnvironmentVariable('ChocolateyToolsLocation', [string]::Empty, $scope)
+    $agentService = Get-Service -Name chocolatey-agent -ErrorAction SilentlyContinue
+    if ($agentService -and $agentService.Status -eq 'Running') {
+        $agentService.Stop()
+    }
+
+    Remove-Item -Path $env:ChocolateyInstall -Recurse -Force
+
+    'ChocolateyInstall', 'ChocolateyLastPathUpdate' | ForEach-Object {
+        foreach ($scope in 'User', 'Machine') {
+            [Environment]::SetEnvironmentVariable($_, [string]::Empty, $scope)
+        }
+    }
+
+    $machineKey.Close()
+    $userKey.Close()
+
+    if ($env:ChocolateyToolsLocation -and (Test-Path $env:ChocolateyToolsLocation)) {
+        Remove-Item -Path $env:ChocolateyToolsLocation -Recurse -Force
+    }
+
+    foreach ($scope in 'User', 'Machine') {
+        [Environment]::SetEnvironmentVariable('ChocolateyToolsLocation', [string]::Empty, $scope)
+    }
 }
 
 Optimize-Volume -DriveLetter C -Analyze -Defrag
